+++ sertop/sertop.ml
@@ -19,7 +19,7 @@
 open Cmdliner
 
 let sertop_version = Ser_version.ser_git_version
-let sertop printer print0 debug lheader coq_path ml_path no_init lp1 lp2 std_impl async deep_edits async_workers omit_loc omit_att exn_on_opaque =
+let sertop printer print0 debug lheader coq_path ml_path no_init lp1 lp2 _std_impl async deep_edits async_workers omit_loc omit_att exn_on_opaque =
 
   let open  Sertop_init         in
   let open! Sertop_sexp         in
@@ -27,7 +27,7 @@ let sertop printer print0 debug lheader coq_path ml_path no_init lp1 lp2 std_imp
   let options = Serlib.Serlib_init.{ omit_loc; omit_att; exn_on_opaque } in
   Serlib.Serlib_init.init ~options;
 
-  let loadpath = Serapi_paths.coq_loadpath_default ~implicit:std_impl ~coq_path @
+  let loadpath = Serapi_paths.coq_loadpath_default ~implicit:true ~coq_path @
                  ml_path @ lp1 @ lp2 in
 
   ser_loop